[REF] web_tour: lazy load hoot-dom for tours
In this commit, we create bundle to lazy load hoot-dom only when a tour
is started to reduce assets weight and impact of web_tour addon.

With this commit, we remove hoot-dom from assets => -100Ko.
And assets of web_tour change from 60Ko to 25Ko.

closes odoo/odoo#226768

X-original-commit: 04d9d218121e39bdb3a2daedeaf643fb12ad1b35
Related: odoo/enterprise#94527
