+++ addons/web_tour/__manifest__.py
@@ -17,16 +17,48 @@ Odoo Web tours.
     ],
     'assets': {
         'web.assets_backend': [
-            'web_tour/static/src/**/*',
-            'web/static/lib/hoot-dom/**/*',
+            'web_tour/static/src/scss/**/*',
+            'web_tour/static/src/js/tour_pointer/**/*',
+            'web_tour/static/src/js/utils/**/*',
+            'web_tour/static/src/js/tour_state.js',
+            'web_tour/static/src/js/tour_service.js',
+            'web_tour/static/src/js/tour_recorder/tour_recorder_state.js',
+            'web_tour/static/src/tour_utils.js',
+            'web_tour/static/src/js/onboarding_item.xml',
         ],
         'web.assets_frontend': [
-            'web_tour/static/src/js/**/*',
+            'web_tour/static/src/scss/**/*',
+            'web_tour/static/src/js/tour_pointer/**/*',
+            'web_tour/static/src/js/utils/**/*',
+            'web_tour/static/src/js/tour_state.js',
+            'web_tour/static/src/js/tour_service.js',
+            'web_tour/static/src/js/tour_recorder/tour_recorder_state.js',
             'web_tour/static/src/tour_utils.js',
-            'web/static/lib/hoot-dom/**/*',
+            'web_tour/static/src/js/onboarding_item.xml',
         ],
         'web.assets_unit_tests': [
+            ('include', 'web_tour.recorder'),
             'web_tour/static/tests/*.test.js',
+        ],
+        'web_tour.common': [
+            'web/static/lib/hoot-dom/**/*',
+            'web_tour/static/src/js/tour_step.js',
+        ],
+        'web_tour.interactive': [
+            ('include', 'web_tour.common'),
+            'web_tour/static/src/js/tour_interactive/**/*',
+            'web_tour/static/src/views/**/*',
+            'web_tour/static/src/widgets/**/*',
+        ],
+        'web_tour.automatic': [
+            ('include', 'web_tour.common'),
+            'web_tour/static/src/js/tour_automatic/**/*',
+        ],
+        'web_tour.recorder': [
+            ('include', 'web_tour.common'),
+            'web_tour/static/src/js/tour_recorder/**/*',
+            'web_tour/static/src/views/**/*',
+            'web_tour/static/src/widgets/**/*',
         ],
     },
     'auto_install': True,
