+++ addons/web_tour/static/tests/tour_automatic.test.js
@@ -8,6 +8,7 @@ import {
     makeMockEnv,
     mountWithCleanup,
     patchWithCleanup,
+    preloadBundle,
 } from "@web/../tests/web_test_helpers";
 import { browser } from "@web/core/browser/browser";
 import { Dialog } from "@web/core/dialog/dialog";
@@ -16,6 +17,8 @@ import { registry } from "@web/core/registry";
 import { useService } from "@web/core/utils/hooks";
 
 describe.current.tags("desktop");
+
+preloadBundle("web_tour.automatic");
 
 const tourRegistry = registry.category("web_tour.tours");
 let macro;
