+++ addons/web_tour/static/tests/tour_interactive.test.js
@@ -13,6 +13,7 @@ import {
     models,
     fields,
     defineModels,
+    preloadBundle,
 } from "@web/../tests/web_test_helpers";
 import { browser } from "@web/core/browser/browser";
 import { Dialog } from "@web/core/dialog/dialog";
@@ -21,6 +22,8 @@ import { session } from "@web/session";
 import { WebClient } from "@web/webclient/webclient";
 
 describe.current.tags("desktop");
+
+preloadBundle("web_tour.interactive");
 
 class Partner extends models.Model {
     _name = "partner";
