+++ addons/web_tour/static/tests/tour_recorder.test.js
@@ -7,20 +7,23 @@ import {
     mountWithCleanup,
     onRpc,
     patchWithCleanup,
+    preloadBundle,
     serverState,
 } from "@web/../tests/web_test_helpers";
 import { AutoComplete } from "@web/core/autocomplete/autocomplete";
 import { browser } from "@web/core/browser/browser";
+import { TourRecorder } from "@web_tour/js/tour_recorder/tour_recorder";
 import {
     TOUR_RECORDER_ACTIVE_LOCAL_STORAGE_KEY,
-    TourRecorder,
-} from "@web_tour/js/tour_recorder/tour_recorder";
+    tourRecorderState,
+} from "@web_tour/js/tour_recorder/tour_recorder_state";
 import { Component, xml } from "@odoo/owl";
 import { useAutofocus } from "@web/core/utils/hooks";
 import { WebClient } from "@web/webclient/webclient";
-import { tourRecorderState } from "@web_tour/js/tour_recorder/tour_recorder_state";
 
 describe.current.tags("desktop");
+
+preloadBundle("web_tour.recorder");
 
 let tourRecorder;
